별점 등록/조회 기능 구현

+++ src/main/java/com/example/block/domain/mapping/Rating.java
@@ -29,5 +29,5 @@ public class Rating {
     private Contest contest;
 
     @Column(nullable = false)
-    private Float score;
+    private double score;
 }
